trying to make the system less fictional
Basically work to build an end-to-end early version of the compiler using the
most naive strategy imaginable of just histogramming all the colors on a given
line and setting the background color to that. I will then build the program
out to include more complicated fitting heuristics (including probably Genetic
Algorithms!!) as well as other more hand-tuned fitting approaches.